добавляет переключение экранов
